Fixed lost change in Extern.h header.

Conflicts:
	src/CodeGen_GPU_Host.cpp

+++ test/correctness/side_effects.cpp
@@ -32,8 +32,8 @@ extern "C" DLLEXPORT int draw_pixel(int x, int y, int val) {
 
     const char *code = " .:-~*={}&%#@";
 
-    if (val >= strlen(code)) {
-        val = strlen(code)-1;
+    if (val >= static_cast<int>(strlen(code))) {
+        val = static_cast<int>(strlen(code))-1;
     }
     printf("%c", code[val]);
     return 0;
